requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20241021
Django==4.2.16
django-bootstrap3==24.3
django-cors-headers==4.5.0
django-crispy-forms==2.3
django-filter==24.3
django-formtools==2.5.1
django-import-export==4.2.0
django-reversion==5.1.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
future==0.18.3
httplib2==0.22.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
reversion==0.2
six==1.16.0
sqlparse==0.5.1
tablib==3.7.0
toml==0.10.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
